http_api: make debug output compatible with bup
Originally, the debug output sent 1:1 the internal database structure.
This never turned out to be useful.
But the raw data can be used to quickly export a match, which turned out to be useful, for example to quickly transfer the match for printing to another computer.
Therefore, output the same format we output to bup.